Feat: workspace enhancements.
Workspaces can now be provided in a simple table rather than specifying
a workspace module (though that remains the default).

Workspaces can now automatically bind keymaps on entry with a maps
table. Done this way so they can eventually be disabled by the workspace
manager as well.

Workspaces can now imply other workspaces. If workspace A implies
workspace B (A=>b) then when A is enabled, B will also be enabled.
Currently B is enabled before A, but this may change in the future.

Condensed activation callbacks into a single activate() callback, having
multiple was redundant and unnecessary.