Fix: project autodetection false positives.
Project autodetection would have false positives when a path was opened
which started with a project directory. This was mainly a problem when
adjacent directories had similar names. This means that
`/path/to/project2` would be detected as `/path/to/proj`.

Now we use fnamemodify() to ensure that we always have a full vaild
path, and never compare *part* of a directory name.