refactor!: changed the api and internals
The major api changes are in the setup function. Now it takes a
workspaces table and a projects table, which setup each submodule
respectively. The individual tables are the same as they were before.
Added functionality for a workspace module back in, that was just nice.
Removed the config module in favor of better code organization.